feat: wrapping on thinner terminals
Moving to "rich" - partially an experiment to investigate the library,
but also the tables support wrapping when the terminal is thinner.

It feels less buggy as well - for me, with tabulate the layout would
sometimes be slightly crazy, and fixed by re-installing the
dependencies.

It's also nice to not have to worry about ANSI escape sequences.

The ASCII style is almost the same as tabulate's.